+++ Backend/src/core/config/app.config.ts
@@ -51,3 +51,8 @@ export const REDIS_CONFIG = {
     port: ENV.REDIS_PORT,
     password: ENV.REDIS_PASSWORD,
 };
+
+export const UPSTASH_REDIS_CONFIG = {
+    url: ENV.UPSTASH_REDIS_REST_URL,
+    token: ENV.UPSTASH_REDIS_REST_TOKEN,
+};